+++ src-py/utils/scanner.py
@@ -1,0 +1,45 @@
+import base64
+from datetime import datetime
+from os import remove
+
+import cv2
+
+
+def preload_model():
+    from deepface import DeepFace
+
+    """
+    Preload to avoid latency during first call
+    """
+    # NOTE: we will hardcode the img
+    img = "/9j/4AAQSkZJRgABAgEASABIAAD/4QViRXhpZgAATU0AKgAAAAgABwESAAMAAAABAAEAAAEaAAUAAAABAAAAYgEbAAUAAAABAAAAagEoAAMAAAABAAIAAAExAAIAAAAdAAAAcgEyAAIAAAAUAAAAj4dpAAQAAAABAAAApAAAANAAAABIAAAAAQAAAEgAAAABQWRvYmUgUGhvdG9zaG9wIENTIE1hY2ludG9zaAAyMDA3OjAyOjE1IDIzOjI5OjQwAAAAA6ABAAMAAAABAAEAAKACAAQAAAABAAAAyKADAAQAAAABAAABBAAAAAAAAAAGAQMAAwAAAAEABgAAARoABQAAAAEAAAEeARsABQAAAAEAAAEmASgAAwAAAAEAAgAAAgEABAAAAAEAAAEuAgIABAAAAAEAAAQsAAAAAAAAAEgAAAABAAAASAAAAAH/2P/gABBKRklGAAECAQBIAEgAAP/tAAxBZG9iZV9DTQAB/+4ADkFkb2JlAGSAAAAAAf/bAIQADAgICAkIDAkJDBELCgsRFQ8MDA8VGBMTFRMTGBEMDAwMDAwRDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAENCwsNDg0QDg4QFA4ODhQUDg4ODhQRDAwMDAwREQwMDAwMDBEMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwM/8AAEQgAoAB7AwEiAAIRAQMRAf/dAAQACP/EAT8AAAEFAQEBAQEBAAAAAAAAAAMAAQIEBQYHCAkKCwEAAQUBAQEBAQEAAAAAAAAAAQACAwQFBgcICQoLEAABBAEDAgQCBQcGCAUDDDMBAAIRAwQhEjEFQVFhEyJxgTIGFJGhsUIjJBVSwWIzNHKC0UMHJZJT8OHxY3M1FqKygyZEk1RkRcKjdDYX0lXiZfKzhMPTdePzRieUpIW0lcTU5PSltcXV5fVWZnaGlqa2xtbm9jdHV2d3h5ent8fX5/cRAAICAQIEBAMEBQYHBwYFNQEAAhEDITESBEFRYXEiEwUygZEUobFCI8FS0fAzJGLhcoKSQ1MVY3M08SUGFqKygwcmNcLSRJNUoxdkRVU2dGXi8rOEw9N14/NGlKSFtJXE1OT0pbXF1eX1VmZ2hpamtsbW5vYnN0dXZ3eHl6e3x//aAAwDAQACEQMRAD8A75jEVrU7WojWpKYhqmGqQapgJKYBqkGqcJQkpjtS2qcJQkphtS2qcJQkpGWqJaiwmISUgLFAtVgtUHNSU1nMUNisuaobUlP/0PSGtRQ1JoUgElKAUgEgE6SlQknSSUsnSSSUpJJJJS0JQnSSUwIUSERMQkpC5qjtRSFGElP/0fTwFIBMApJKUnSCSSlJJJJKUkkkkpSSSSSlJJJJKUmTpJKYkKMKZUUlP//S9RCcJJwkpSSSSSlJJJ0lKSSSSUpJJJJSkydMkpSSSSSlFMnTJKf/0/Uk6ZOkpSSSSSlJ0ydJSkkkklKSSSSUpJJJJSySSSSlJJJJKf/U9STpk4SUpJJJJSk6ZJJS6SSSSlJJJJKUmTpklKSSSSUpMnTJKf/V9RCdRaVJJS6SSSSlJJJJKUkkkkpSSSSSlJJJJKUkkkkpRUU5UUlP/9b04FTBQWlTBSUkTqIKdJS6SSSSlJJJJKUkkkkpSSSSSlJiUiUxKSlFRlIlQ3JKf//X9Ha5FDlVa9Fa5JScFSBQQ5TDklJZSlQBTykpmkoSlKSmaUqEpSkpkSmlNKYlJS8qJcmLlAuSUuXKO5Qc5R3JKf/Q7pr0ZtipNsRWvSU3GvRA9U2vRWvSU2g4KW5Vg9SD/NJSeU8oG9PvSUmlNuQt6bekpKXqJchl6iXpKSFyg56GXqDnpKZueh70N1iHvKSn/9n/7Sg+UGhvdG9zaG9wIDMuMAA4QklNBCUAAAAAABAAAAAAAAAAAAAAAAAAAAAAOEJJTQPqAAAAAB2wPD94bWwgdmVyc2lvbj0iMS4wIiBlbmNvZGluZz0iVVRGLTgiPz4KPCFET0NUWVBFIHBsaXN0IFBVQkxJQyAiLS8vQXBwbGUgQ29tcHV0ZXIvL0RURCBQTElTVCAxLjAvL0VOIiAiaHR0cDovL3d3dy5hcHBsZS5jb20vRFREcy9Qcm9wZXJ0eUxpc3QtMS4wLmR0ZCI+CjxwbGlzdCB2ZXJzaW9uPSIxLjAiPgo8ZGljdD4KCTxrZXk+Y29tLmFwcGxlLnByaW50LlBhZ2VGb3JtYXQuUE1Ib3Jpem9udGFsUmVzPC9rZXk+Cgk8ZGljdD4KCQk8a2V5PmNvbS5hcHBsZS5wcmludC50aWNrZXQuY3JlYXRvcjwva2V5PgoJCTxzdHJpbmc+Y29tLmFwcGxlLnByaW50aW5nbWFuYWdlcjwvc3RyaW5nPgoJCTxrZXk+Y29tLmFwcGxlLnByaW50LnRpY2tldC5pdGVtQXJyYXk8L2tleT4KCQk8YXJyYXk+CgkJCTxkaWN0PgoJCQkJPGtleT5jb20uYXBwbGUucHJpbnQuUGFnZUZvcm1hdC5QTUhvcml6b250YWxSZXM8L2tleT4KCQkJCTxyZWFsPjcyPC9yZWFsPgoJCQkJPGtleT5jb20uYXBwbGUucHJpbnQudGlja2V0LmNsaWVudDwva2V5PgoJCQkJPHN0cmluZz5jb20uYXBwbGUucHJpbnRpbmdtYW5hZ2VyPC9zdHJpbmc+CgkJCQk8a2V5PmNvbS5hcHBsZS5wcmludC50aWNrZXQubW9kRGF0ZTwva2V5PgoJCQkJPGRhdGU+MjAwNy0wMi0xNlQwNDoyNzowMFo8L2RhdGU+CgkJCQk8a2V5PmNvbS5hcHBsZS5wcmludC50aWNrZXQuc3RhdGVGbGFnPC9rZXk+CgkJCQk8aW50ZWdlcj4wPC9pbnRlZ2VyPgoJCQk8L2RpY3Q+CgkJPC9hcnJheT4KCTwvZGljdD4KCTxrZXk+Y29tLmFwcGxlLnByaW50LlBhZ2VGb3JtYXQuUE1PcmllbnRhdGlvbjwva2V5PgoJPGRpY3Q+CgkJPGtleT5jb20uYXBwbGUucHJpbnQudGlja2V0LmNyZWF0b3I8L2tleT4KCQk8c3RyaW5nPmNvbS5hcHBsZS5wcmludGluZ21hbmFnZXI8L3N0cmluZz4KCQk8a2V5PmNvbS5hcHBsZS5wcmludC50aWNrZXQuaXRlbUFycmF5PC9rZXk+CgkJPGFycmF5PgoJCQk8ZGljdD4KCQkJCTxrZXk+Y29tLmFwcGxlLnByaW50LlBhZ2VGb3JtYXQuUE1PcmllbnRhdGlvbjwva2V5PgoJCQkJPGludGVnZXI+MTwvaW50ZWdlcj4KCQkJCTxrZXk+Y29tLmFwcGxlLnByaW50LnRpY2tldC5jbGllbnQ8L2tleT4KCQkJCTxzdHJpbmc+Y29tLmFwcGxlLnByaW50aW5nbWFuYWdlcjwvc3RyaW5nPgoJCQkJPGtleT5jb20uYXBwbGUucHJpbnQudGlja2V0Lm1vZERhdGU8L2tleT4KCQkJCTxkYXRlPjIwMDctMDItMTZUMDQ6Mjc6MDBaPC9kYXRlPgoJCQkJPGtleT5jb20uYXBwbGUucHJpbnQudGlja2V0LnN0YXRlRmxhZzwva2V5PgoJCQkJPGludGVnZXI+MDwvaW50ZWdlcj4KCQkJPC9kaWN0PgoJCTwvYXJyYXk+Cgk8L2RpY3Q+Cgk8a2V5PmNvbS5hcHBsZS5wcmludC5QYWdlRm9ybWF0LlBNU2NhbGluZzwva2V5PgoJPGRpY3Q+CgkJPGtleT5jb20uYXBwbGUucHJpbnQudGlja2V0LmNyZWF0b3I8L2tleT4KCQk8c3RyaW5nPmNvbS5hcHBsZS5wcmludGluZ21hbmFnZXI8L3N0cmluZz4KCQk8a2V5PmNvbS5hcHBsZS5wcmludC50aWNrZXQuaXRlbUFycmF5PC9rZXk+CgkJPGFycmF5PgoJCQk8ZGljdD4KCQkJCTxrZXk+Y29tLmFwcGxlLnByaW50LlBhZ2VGb3JtYXQuUE1TY2FsaW5nPC9rZXk+CgkJCQk8cmVhbD4xPC9yZWFsPgoJCQkJPGtleT5jb20uYXBwbGUucHJpbnQudGlja2V0LmNsaWVudDwva2V5PgoJCQkJPHN0cmluZz5jb20uYXBwbGUucHJpbnRpbmdtYW5hZ2VyPC9zdHJpbmc+CgkJCQk8a2V5PmNvbS5hcHBsZS5wcmludC50aWNrZXQubW9kRGF0ZTwva2V5PgoJCQkJPGRhdGU+MjAwNy0wMi0xNlQwNDoyNzowMFo8L2RhdGU+CgkJCQk8a2V5PmNvbS5hcHBsZS5wcmludC50aWNrZXQuc3RhdGVGbGFnPC9rZXk+CgkJCQk8aW50ZWdlcj4wPC9pbnRlZ2VyPgoJCQk8L2RpY3Q+CgkJPC9hcnJheT4KCTwvZGljdD4KCTxrZXk+Y29tLmFwcGxlLnByaW50LlBhZ2VGb3JtYXQuUE1WZXJ0aWNhbFJlczwva2V5PgoJPGRpY3Q+CgkJPGtleT5jb20uYXBwbGUucHJpbnQudGlja2V0LmNyZWF0b3I8L2tleT4KCQk8c3RyaW5nPmNvbS5hcHBsZS5wcmludGluZ21hbmFnZXI8L3N0cmluZz4KCQk8a2V5PmNvbS5hcHBsZS5wcmludC50aWNrZXQuaXRlbUFycmF5PC9rZXk+CgkJPGFycmF5PgoJCQk8ZGljdD4KCQkJCTxrZXk+Y29tLmFwcGxlLnByaW50LlBhZ2VGb3JtYXQuUE1WZXJ0aWNhbFJlczwva2V5PgoJCQkJPHJlYWw+NzI8L3JlYWw+CgkJCQk8a2V5PmNvbS5hcHBsZS5wcmludC50aWNrZXQuY2xpZW50PC9rZXk+CgkJCQk8c3RyaW5nPmNvbS5hcHBsZS5wcmludGluZ21hbmFnZXI8L3N0cmluZz4KCQkJCTxrZXk+Y29tLmFwcGxlLnByaW50LnRpY2tldC5tb2REYXRlPC9rZXk+CgkJCQk8ZGF0ZT4yMDA3LTAyLTE2VDA0OjI3OjAwWjwvZGF0ZT4KCQkJCTxrZXk+Y29tLmFwcGxlLnByaW50LnRpY2tldC5zdGF0ZUZsYWc8L2tleT4KCQkJCTxpbnRlZ2VyPjA8L2ludGVnZXI+CgkJCTwvZGljdD4KCQk8L2FycmF5PgoJPC9kaWN0PgoJPGtleT5jb20uYXBwbGUucHJpbnQuUGFnZUZvcm1hdC5QTVZlcnRpY2FsU2NhbGluZzwva2V5PgoJPGRpY3Q+CgkJPGtleT5jb20uYXBwbGUucHJpbnQudGlja2V0LmNyZWF0b3I8L2tleT4KCQk8c3RyaW5nPmNvbS5hcHBsZS5wcmludGluZ21hbmFnZXI8L3N0cmluZz4KCQk8a2V5PmNvbS5hcHBsZS5wcmludC50aWNrZXQuaXRlbUFycmF5PC9rZXk+CgkJPGFycmF5PgoJCQk8ZGljdD4KCQkJCTxrZXk+Y29tLmFwcGxlLnByaW50LlBhZ2VGb3JtYXQuUE1WZXJ0aWNhbFNjYWxpbmc8L2tleT4KCQkJCTxyZWFsPjE8L3JlYWw+CgkJCQk8a2V5PmNvbS5hcHBsZS5wcmludC50aWNrZXQuY2xpZW50PC9rZXk+CgkJCQk8c3RyaW5nPmNvbS5hcHBsZS5wcmludGluZ21hbmFnZXI8L3N0cmluZz4KCQkJCTxrZXk+Y29tLmFwcGxlLnByaW50LnRpY2tldC5tb2REYXRlPC9rZXk+CgkJCQk8ZGF0ZT4yMDA3LTAyLTE2VDA0OjI3OjAwWjwvZGF0ZT4KCQkJCTxrZXk+Y29tLmFwcGxlLnByaW50LnRpY2tldC5zdGF0ZUZsYWc8L2tleT4KCQkJCTxpbnRlZ2VyPjA8L2ludGVnZXI+CgkJCTwvZGljdD4KCQk8L2FycmF5PgoJPC9kaWN0PgoJPGtleT5jb20uYXBwbGUucHJpbnQuc3ViVGlja2V0LnBhcGVyX2luZm9fdGlja2V0PC9rZXk+Cgk8ZGljdD4KCQk8a2V5PmNvbS5hcHBsZS5wcmludC5QYWdlRm9ybWF0LlBNQWRqdXN0ZWRQYWdlUmVjdDwva2V5PgoJCTxkaWN0PgoJCQk8a2V5PmNvbS5hcHBsZS5wcmludC50aWNrZXQuY3JlYXRvcjwva2V5PgoJCQk8c3RyaW5nPmNvbS5hcHBsZS5wcmludGluZ21hbmFnZXI8L3N0cmluZz4KCQkJPGtleT5jb20uYXBwbGUucHJpbnQudGlja2V0Lml0ZW1BcnJheTwva2V5PgoJCQk8YXJyYXk+CgkJCQk8ZGljdD4KCQkJCQk8a2V5PmNvbS5hcHBsZS5wcmludC5QYWdlRm9ybWF0LlBNQWRqdXN0ZWRQYWdlUmVjdDwva2V5PgoJCQkJCTxhcnJheT4KCQkJCQkJPHJlYWw+MC4wPC9yZWFsPgoJCQkJCQk8cmVhbD4wLjA8L3JlYWw+CgkJCQkJCTxyZWFsPjczNDwvcmVhbD4KCQkJCQkJPHJlYWw+NTc2PC9yZWFsPgoJCQkJCTwvYXJyYXk+CgkJCQkJPGtleT5jb20uYXBwbGUucHJpbnQudGlja2V0LmNsaWVudDwva2V5PgoJCQkJCTxzdHJpbmc+Y29tLmFwcGxlLnByaW50aW5nbWFuYWdlcjwvc3RyaW5nPgoJCQkJCTxrZXk+Y29tLmFwcGxlLnByaW50LnRpY2tldC5tb2REYXRlPC9rZXk+CgkJCQkJPGRhdGU+MjAwNy0wMi0xNlQwNDoyNzowMFo8L2RhdGU+CgkJCQkJPGtleT5jb20uYXBwbGUucHJpbnQudGlja2V0LnN0YXRlRmxhZzwva2V5PgoJCQkJCTxpbnRlZ2VyPjA8L2ludGVnZXI+CgkJCQk8L2RpY3Q+CgkJCTwvYXJyYXk+CgkJPC9kaWN0PgoJCTxrZXk+Y29tLmFwcGxlLnByaW50LlBhZ2VGb3JtYXQuUE1BZGp1c3RlZFBhcGVyUmVjdDwva2V5PgoJCTxkaWN0PgoJCQk8a2V5PmNvbS5hcHBsZS5wcmludC50aWNrZXQuY3JlYXRvcjwva2V5PgoJCQk8c3RyaW5nPmNvbS5hcHBsZS5wcmludGluZ21hbmFnZXI8L3N0cmluZz4KCQkJPGtleT5jb20uYXBwbGUucHJpbnQudGlja2V0Lml0ZW1BcnJheTwva2V5PgoJCQk8YXJyYXk+CgkJCQk8ZGljdD4KCQkJCQk8a2V5PmNvbS5hcHBsZS5wcmludC5QYWdlRm9ybWF0LlBNQWRqdXN0ZWRQYXBlclJlY3Q8L2tleT4KCQkJCQk8YXJyYXk+CgkJCQkJCTxyZWFsPi0xODwvcmVhbD4KCQkJCQkJPHJlYWw+LTE4PC9yZWFsPgoJCQkJCQk8cmVhbD43NzQ8L3JlYWw+CgkJCQkJCTxyZWFsPjU5NDwvcmVhbD4KCQkJCQk8L2FycmF5PgoJCQkJCTxrZXk+Y29tLmFwcGxlLnByaW50LnRpY2tldC5jbGllbnQ8L2tleT4KCQkJCQk8c3RyaW5nPmNvbS5hcHBsZS5wcmludGluZ21hbmFnZXI8L3N0cmluZz4KCQkJCQk8a2V5PmNvbS5hcHBsZS5wcmludC50aWNrZXQubW9kRGF0ZTwva2V5PgoJCQkJCTxkYXRlPjIwMDctMDItMTZUMDQ6Mjc6MDBaPC9kYXRlPgoJCQkJCTxrZXk+Y29tLmFwcGxlLnByaW50LnRpY2tldC5zdGF0ZUZsYWc8L2tleT4KCQkJCQk8aW50ZWdlcj4wPC9pbnRlZ2VyPgoJCQkJPC9kaWN0PgoJCQk8L2FycmF5PgoJCTwvZGljdD4KCQk8a2V5PmNvbS5hcHBsZS5wcmludC5QYXBlckluZm8uUE1QYXBlck5hbWU8L2tleT4KCQk8ZGljdD4KCQkJPGtleT5jb20uYXBwbGUucHJpbnQudGlja2V0LmNyZWF0b3I8L2tleT4KCQkJPHN0cmluZz5jb20uYXBwbGUucHJpbnQucG0uUG9zdFNjcmlwdDwvc3RyaW5nPgoJCQk8a2V5PmNvbS5hcHBsZS5wcmludC50aWNrZXQuaXRlbUFycmF5PC9rZXk+CgkJCTxhcnJheT4KCQkJCTxkaWN0PgoJCQkJCTxrZXk+Y29tLmFwcGxlLnByaW50LlBhcGVySW5mby5QTVBhcGVyTmFtZTwva2V5PgoJCQkJCTxzdHJpbmc+bmEtbGV0dGVyPC9zdHJpbmc+CgkJCQkJPGtleT5jb20uYXBwbGUucHJpbnQudGlja2V0LmNsaWVudDwva2V5PgoJCQkJCTxzdHJpbmc+Y29tLmFwcGxlLnByaW50LnBtLlBvc3RTY3JpcHQ8L3N0cmluZz4KCQkJCQk8a2V5PmNvbS5hcHBsZS5wcmludC50aWNrZXQubW9kRGF0ZTwva2V5PgoJCQkJCTxkYXRlPjIwMDMtMDctMDFUMTc6NDk6MzZaPC9kYXRlPgoJCQkJCTxrZXk+Y29tLmFwcGxlLnByaW50LnRpY2tldC5zdGF0ZUZsYWc8L2tleT4KCQkJCQk8aW50ZWdlcj4xPC9pbnRlZ2VyPgoJCQkJPC9kaWN0PgoJCQk8L2FycmF5PgoJCTwvZGljdD4KCQk8a2V5PmNvbS5hcHBsZS5wcmludC5QYXBlckluZm8uUE1VbmFkanVzdGVkUGFnZVJlY3Q8L2tleT4KCQk8ZGljdD4KCQkJPGtleT5jb20uYXBwbGUucHJpbnQudGlja2V0LmNyZWF0b3I8L2tleT4KCQkJPHN0cmluZz5jb20uYXBwbGUucHJpbnQucG0uUG9zdFNjcmlwdDwvc3RyaW5nPgoJCQk8a2V5PmNvbS5hcHBsZS5wcmludC50aWNrZXQuaXRlbUFycmF5PC9rZXk+CgkJCTxhcnJheT4KCQkJCTxkaWN0PgoJCQkJCTxrZXk+Y29tLmFwcGxlLnByaW50LlBhcGVySW5mby5QTVVuYWRqdXN0ZWRQYWdlUmVjdDwva2V5PgoJCQkJCTxhcnJheT4KCQkJCQkJPHJlYWw+MC4wPC9yZWFsPgoJCQkJCQk8cmVhbD4wLjA8L3JlYWw+CgkJCQkJCTxyZWFsPjczNDwvcmVhbD4KCQkJCQkJPHJlYWw+NTc2PC9yZWFsPgoJCQkJCTwvYXJyYXk+CgkJCQkJPGtleT5jb20uYXBwbGUucHJpbnQudGlja2V0LmNsaWVudDwva2V5PgoJCQkJCTxzdHJpbmc+Y29tLmFwcGxlLnByaW50aW5nbWFuYWdlcjwvc3RyaW5nPgoJCQkJCTxrZXk+Y29tLmFwcGxlLnByaW50LnRpY2tldC5tb2REYXRlPC9rZXk+CgkJCQkJPGRhdGU+MjAwNy0wMi0xNlQwNDoyNzowMFo8L2RhdGU+CgkJCQkJPGtleT5jb20uYXBwbGUucHJpbnQudGlja2V0LnN0YXRlRmxhZzwva2V5PgoJCQkJCTxpbnRlZ2VyPjA8L2ludGVnZXI+CgkJCQk8L2RpY3Q+CgkJCTwvYXJyYXk+CgkJPC9kaWN0PgoJCTxrZXk+Y29tLmFwcGxlLnByaW50LlBhcGVySW5mby5QTVVuYWRqdXN0ZWRQYXBlclJlY3Q8L2tleT4KCQk8ZGljdD4KCQkJPGtleT5jb20uYXBwbGUucHJpbnQudGlja2V0LmNyZWF0b3I8L2tleT4KCQkJPHN0cmluZz5jb20uYXBwbGUucHJpbnQucG0uUG9zdFNjcmlwdDwvc3RyaW5nPgoJCQk8a2V5PmNvbS5hcHBsZS5wcmludC50aWNrZXQuaXRlbUFycmF5PC9rZXk+CgkJCTxhcnJheT4KCQkJCTxkaWN0PgoJCQkJCTxrZXk+Y29tLmFwcGxlLnByaW50LlBhcGVySW5mby5QTVVuYWRqdXN0ZWRQYXBlclJlY3Q8L2tleT4KCQkJCQk8YXJyYXk+CgkJCQkJCTxyZWFsPi0xODwvcmVhbD4KCQkJCQkJPHJlYWw+LTE4PC9yZWFsPgoJCQkJCQk8cmVhbD43NzQ8L3JlYWw+CgkJCQkJCTxyZWFsPjU5NDwvcmVhbD4KCQkJCQk8L2FycmF5PgoJCQkJCTxrZXk+Y29tLmFwcGxlLnByaW50LnRpY2tldC5jbGllbnQ8L2tleT4KCQkJCQk8c3RyaW5nPmNvbS5hcHBsZS5wcmludGluZ21hbmFnZXI8L3N0cmluZz4KCQkJCQk8a2V5PmNvbS5hcHBsZS5wcmludC50aWNrZXQubW9kRGF0ZTwva2V5PgoJCQkJCTxkYXRlPjIwMDctMDItMTZUMDQ6Mjc6MDBaPC9kYXRlPgoJCQkJCTxrZXk+Y29tLmFwcGxlLnByaW50LnRpY2tldC5zdGF0ZUZsYWc8L2tleT4KCQkJCQk8aW50ZWdlcj4wPC9pbnRlZ2VyPgoJCQkJPC9kaWN0PgoJCQk8L2FycmF5PgoJCTwvZGljdD4KCQk8a2V5PmNvbS5hcHBsZS5wcmludC5QYXBlckluZm8ucHBkLlBNUGFwZXJOYW1lPC9rZXk+CgkJPGRpY3Q+CgkJCTxrZXk+Y29tLmFwcGxlLnByaW50LnRpY2tldC5jcmVhdG9yPC9rZXk+CgkJCTxzdHJpbmc+Y29tLmFwcGxlLnByaW50LnBtLlBvc3RTY3JpcHQ8L3N0cmluZz4KCQkJPGtleT5jb20uYXBwbGUucHJpbnQudGlja2V0Lml0ZW1BcnJheTwva2V5PgoJCQk8YXJyYXk+CgkJCQk8ZGljdD4KCQkJCQk8a2V5PmNvbS5hcHBsZS5wcmludC5QYXBlckluZm8ucHBkLlBNUGFwZXJOYW1lPC9rZXk+CgkJCQkJPHN0cmluZz5VUyBMZXR0ZXI8L3N0cmluZz4KCQkJCQk8a2V5PmNvbS5hcHBsZS5wcmludC50aWNrZXQuY2xpZW50PC9rZXk+CgkJCQkJPHN0cmluZz5jb20uYXBwbGUucHJpbnQucG0uUG9zdFNjcmlwdDwvc3RyaW5nPgoJCQkJCTxrZXk+Y29tLmFwcGxlLnByaW50LnRpY2tldC5tb2REYXRlPC9rZXk+CgkJCQkJPGRhdGU+MjAwMy0wNy0wMVQxNzo0OTozNlo8L2RhdGU+CgkJCQkJPGtleT5jb20uYXBwbGUucHJpbnQudGlja2V0LnN0YXRlRmxhZzwva2V5PgoJCQkJCTxpbnRlZ2VyPjE8L2ludGVnZXI+CgkJCQk8L2RpY3Q+CgkJCTwvYXJyYXk+CgkJPC9kaWN0PgoJCTxrZXk+Y29tLmFwcGxlLnByaW50LnRpY2tldC5BUElWZXJzaW9uPC9rZXk+CgkJPHN0cmluZz4wMC4yMDwvc3RyaW5nPgoJCTxrZXk+Y29tLmFwcGxlLnByaW50LnRpY2tldC5wcml2YXRlTG9jazwva2V5PgoJCTxmYWxzZS8+CgkJPGtleT5jb20uYXBwbGUucHJpbnQudGlja2V0LnR5cGU8L2tleT4KCQk8c3RyaW5nPmNvbS5hcHBsZS5wcmludC5QYXBlckluZm9UaWNrZXQ8L3N0cmluZz4KCTwvZGljdD4KCTxrZXk+Y29tLmFwcGxlLnByaW50LnRpY2tldC5BUElWZXJzaW9uPC9rZXk+Cgk8c3RyaW5nPjAwLjIwPC9zdHJpbmc+Cgk8a2V5PmNvbS5hcHBsZS5wcmludC50aWNrZXQucHJpdmF0ZUxvY2s8L2tleT4KCTxmYWxzZS8+Cgk8a2V5PmNvbS5hcHBsZS5wcmludC50aWNrZXQudHlwZTwva2V5PgoJPHN0cmluZz5jb20uYXBwbGUucHJpbnQuUGFnZUZvcm1hdFRpY2tldDwvc3RyaW5nPgo8L2RpY3Q+CjwvcGxpc3Q+CjhCSU0D6QAAAAAAeAADAAAASABIAAAAAALeAkD/7v/uAwYCUgNnBSgD/AACAAAASABIAAAAAALYAigAAQAAAGQAAAABAAMDAwAAAAF//wABAAEAAAAAAAAAAAAAAABoCAAZAZAAAAAAACAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAADhCSU0D7QAAAAAAEABIAAAAAQABAEgAAAABAAE4QklNBCYAAAAAAA4AAAAAAAAAAAAAP4AAADhCSU0EDQAAAAAABAAAAHg4QklNBBkAAAAAAAQAAAAeOEJJTQPzAAAAAAAJAAAAAAAAAAABADhCSU0ECgAAAAAAAQAAOEJJTScQAAAAAAAKAAEAAAAAAAAAAThCSU0D9QAAAAAASAAvZmYAAQBsZmYABgAAAAAAAQAvZmYAAQChmZoABgAAAAAAAQAyAAAAAQBaAAAABgAAAAAAAQA1AAAAAQAtAAAABgAAAAAAAThCSU0D+AAAAAAAcAAA/////////////////////////////wPoAAAAAP////////////////////////////8D6AAAAAD/////////////////////////////A+gAAAAA/////////////////////////////wPoAAA4QklNBAgAAAAAABAAAAABAAACQAAAAkAAAAAAOEJJTQQeAAAAAAAEAAAAADhCSU0EGgAAAAADSQAAAAYAAAAAAAAAAAAAAQQAAADIAAAACgBVAG4AdABpAHQAbABlAGQALQAxAAAAAQAAAAAAAAAAAAAAAAAAAAAAAAABAAAAAAAAAAAAAADIAAABBAAAAAAAAAAAAAAAAAAAAAABAAAAAAAAAAAAAAAAAAAAAAAAABAAAAABAAAAAAAAbnVsbAAAAAIAAAAGYm91bmRzT2JqYwAAAAEAAAAAAABSY3QxAAAABAAAAABUb3AgbG9uZwAAAAAAAAAATGVmdGxvbmcAAAAAAAAAAEJ0b21sb25nAAABBAAAAABSZ2h0bG9uZwAAAMgAAAAGc2xpY2VzVmxMcwAAAAFPYmpjAAAAAQAAAAAABXNsaWNlAAAAEgAAAAdzbGljZUlEbG9uZwAAAAAAAAAHZ3JvdXBJRGxvbmcAAAAAAAAABm9yaWdpbmVudW0AAAAMRVNsaWNlT3JpZ2luAAAADWF1dG9HZW5lcmF0ZWQAAAAAVHlwZWVudW0AAAAKRVNsaWNlVHlwZQAAAABJbWcgAAAABmJvdW5kc09iamMAAAABAAAAAAAAUmN0MQAAAAQAAAAAVG9wIGxvbmcAAAAAAAAAAExlZnRsb25nAAAAAAAAAABCdG9tbG9uZwAAAQQAAAAAUmdodGxvbmcAAADIAAAAA3VybFRFWFQAAAABAAAAAAAAbnVsbFRFWFQAAAABAAAAAAAATXNnZVRFWFQAAAABAAAAAAAGYWx0VGFnVEVYVAAAAAEAAAAAAA5jZWxsVGV4dElzSFRNTGJvb2wBAAAACGNlbGxUZXh0VEVYVAAAAAEAAAAAAAlob3J6QWxpZ25lbnVtAAAAD0VTbGljZUhvcnpBbGlnbgAAAAdkZWZhdWx0AAAACXZlcnRBbGlnbmVudW0AAAAPRVNsaWNlVmVydEFsaWduAAAAB2RlZmF1bHQAAAALYmdDb2xvclR5cGVlbnVtAAAAEUVTbGljZUJHQ29sb3JUeXBlAAAAAE5vbmUAAAAJdG9wT3V0c2V0bG9uZwAAAAAAAAAKbGVmdE91dHNldGxvbmcAAAAAAAAADGJvdHRvbU91dHNldGxvbmcAAAAAAAAAC3JpZ2h0T3V0c2V0bG9uZwAAAAAAOEJJTQQoAAAAAAAMAAAAAT/wAAAAAAAAOEJJTQQUAAAAAAAEAAAAAThCSU0EDAAAAAAESAAAAAEAAAB7AAAAoAAAAXQAAOiAAAAELAAYAAH/2P/gABBKRklGAAECAQBIAEgAAP/tAAxBZG9iZV9DTQAB/+4ADkFkb2JlAGSAAAAAAf/bAIQADAgICAkIDAkJDBELCgsRFQ8MDA8VGBMTFRMTGBEMDAwMDAwRDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAENCwsNDg0QDg4QFA4ODhQUDg4ODhQRDAwMDAwREQwMDAwMDBEMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwM/8AAEQgAoAB7AwEiAAIRAQMRAf/dAAQACP/EAT8AAAEFAQEBAQEBAAAAAAAAAAMAAQIEBQYHCAkKCwEAAQUBAQEBAQEAAAAAAAAAAQACAwQFBgcICQoLEAABBAEDAgQCBQcGCAUDDDMBAAIRAwQhEjEFQVFhEyJxgTIGFJGhsUIjJBVSwWIzNHKC0UMHJZJT8OHxY3M1FqKygyZEk1RkRcKjdDYX0lXiZfKzhMPTdePzRieUpIW0lcTU5PSltcXV5fVWZnaGlqa2xtbm9jdHV2d3h5ent8fX5/cRAAICAQIEBAMEBQYHBwYFNQEAAhEDITESBEFRYXEiEwUygZEUobFCI8FS0fAzJGLhcoKSQ1MVY3M08SUGFqKygwcmNcLSRJNUoxdkRVU2dGXi8rOEw9N14/NGlKSFtJXE1OT0pbXF1eX1VmZ2hpamtsbW5vYnN0dXZ3eHl6e3x//aAAwDAQACEQMRAD8A75jEVrU7WojWpKYhqmGqQapgJKYBqkGqcJQkpjtS2qcJQkphtS2qcJQkpGWqJaiwmISUgLFAtVgtUHNSU1nMUNisuaobUlP/0PSGtRQ1JoUgElKAUgEgE6SlQknSSUsnSSSUpJJJJS0JQnSSUwIUSERMQkpC5qjtRSFGElP/0fTwFIBMApJKUnSCSSlJJJJKUkkkkpSSSSSlJJJJKUmTpJKYkKMKZUUlP//S9RCcJJwkpSSSSSlJJJ0lKSSSSUpJJJJSkydMkpSSSSSlFMnTJKf/0/Uk6ZOkpSSSSSlJ0ydJSkkkklKSSSSUpJJJJSySSSSlJJJJKf/U9STpk4SUpJJJJSk6ZJJS6SSSSlJJJJKUmTpklKSSSSUpMnTJKf/V9RCdRaVJJS6SSSSlJJJJKUkkkkpSSSSSlJJJJKUkkkkpRUU5UUlP/9b04FTBQWlTBSUkTqIKdJS6SSSSlJJJJKUkkkkpSSSSSlJiUiUxKSlFRlIlQ3JKf//X9Ha5FDlVa9Fa5JScFSBQQ5TDklJZSlQBTykpmkoSlKSmaUqEpSkpkSmlNKYlJS8qJcmLlAuSUuXKO5Qc5R3JKf/Q7pr0ZtipNsRWvSU3GvRA9U2vRWvSU2g4KW5Vg9SD/NJSeU8oG9PvSUmlNuQt6bekpKXqJchl6iXpKSFyg56GXqDnpKZueh70N1iHvKSn/9k4QklNBCEAAAAAAFMAAAABAQAAAA8AQQBkAG8AYgBlACAAUABoAG8AdABvAHMAaABvAHAAAAASAEEAZABvAGIAZQAgAFAAaABvAHQAbwBzAGgAbwBwACAAQwBTAAAAAQA4QklNBAYAAAAAAAcACAAAAAEBAP/hF/xodHRwOi8vbnMuYWRvYmUuY29tL3hhcC8xLjAvADw/eHBhY2tldCBiZWdpbj0n77u/JyBpZD0nVzVNME1wQ2VoaUh6cmVTek5UY3prYzlkJz8+Cjx4OnhtcG1ldGEgeG1sbnM6eD0nYWRvYmU6bnM6bWV0YS8nIHg6eG1wdGs9J1hNUCB0b29sa2l0IDMuMC0yOCwgZnJhbWV3b3JrIDEuNic+CjxyZGY6UkRGIHhtbG5zOnJkZj0naHR0cDovL3d3dy53My5vcmcvMTk5OS8wMi8yMi1yZGYtc3ludGF4LW5zIycgeG1sbnM6aVg9J2h0dHA6Ly9ucy5hZG9iZS5jb20vaVgvMS4wLyc+CgogPHJkZjpEZXNjcmlwdGlvbiByZGY6YWJvdXQ9J3V1aWQ6NTAxYWViNTktYmVkZS0xMWRiLTk4NjEtOTliODAyMjQ0ZDIyJwogIHhtbG5zOmV4aWY9J2h0dHA6Ly9ucy5hZG9iZS5jb20vZXhpZi8xLjAvJz4KICA8ZXhpZjpDb2xvclNwYWNlPjE8L2V4aWY6Q29sb3JTcGFjZT4KICA8ZXhpZjpQaXhlbFhEaW1lbnNpb24+MjAwPC9leGlmOlBpeGVsWERpbWVuc2lvbj4KICA8ZXhpZjpQaXhlbFlEaW1lbnNpb24+MjYwPC9leGlmOlBpeGVsWURpbWVuc2lvbj4KIDwvcmRmOkRlc2NyaXB0aW9uPgoKIDxyZGY6RGVzY3JpcHRpb24gcmRmOmFib3V0PSd1dWlkOjUwMWFlYjU5LWJlZGUtMTFkYi05ODYxLTk5YjgwMjI0NGQyMicKICB4bWxuczpwZGY9J2h0dHA6Ly9ucy5hZG9iZS5jb20vcGRmLzEuMy8nPgogPC9yZGY6RGVzY3JpcHRpb24+CgogPHJkZjpEZXNjcmlwdGlvbiByZGY6YWJvdXQ9J3V1aWQ6NTAxYWViNTktYmVkZS0xMWRiLTk4NjEtOTliODAyMjQ0ZDIyJwogIHhtbG5zOnBob3Rvc2hvcD0naHR0cDovL25zLmFkb2JlLmNvbS9waG90b3Nob3AvMS4wLyc+CiAgPHBob3Rvc2hvcDpIaXN0b3J5PjwvcGhvdG9zaG9wOkhpc3Rvcnk+CiA8L3JkZjpEZXNjcmlwdGlvbj4KCiA8cmRmOkRlc2NyaXB0aW9uIHJkZjphYm91dD0ndXVpZDo1MDFhZWI1OS1iZWRlLTExZGItOTg2MS05OWI4MDIyNDRkMjInCiAgeG1sbnM6dGlmZj0naHR0cDovL25zLmFkb2JlLmNvbS90aWZmLzEuMC8nPgogIDx0aWZmOk9yaWVudGF0aW9uPjE8L3RpZmY6T3JpZW50YXRpb24+CiAgPHRpZmY6WFJlc29sdXRpb24+NzIvMTwvdGlmZjpYUmVzb2x1dGlvbj4KICA8dGlmZjpZUmVzb2x1dGlvbj43Mi8xPC90aWZmOllSZXNvbHV0aW9uPgogIDx0aWZmOlJlc29sdXRpb25Vbml0PjI8L3RpZmY6UmVzb2x1dGlvblVuaXQ+CiA8L3JkZjpEZXNjcmlwdGlvbj4KCiA8cmRmOkRlc2NyaXB0aW9uIHJkZjphYm91dD0ndXVpZDo1MDFhZWI1OS1iZWRlLTExZGItOTg2MS05OWI4MDIyNDRkMjInCiAgeG1sbnM6eGFwPSdodHRwOi8vbnMuYWRvYmUuY29tL3hhcC8xLjAvJz4KICA8eGFwOkNyZWF0ZURhdGU+MjAwNy0wMi0xNVQyMzoyOTo0MC0wNTowMDwveGFwOkNyZWF0ZURhdGU+CiAgPHhhcDpNb2RpZnlEYXRlPjIwMDctMDItMTVUMjM6Mjk6NDAtMDU6MDA8L3hhcDpNb2RpZnlEYXRlPgogIDx4YXA6TWV0YWRhdGFEYXRlPjIwMDctMDItMTVUMjM6Mjk6NDAtMDU6MDA8L3hhcDpNZXRhZGF0YURhdGU+CiAgPHhhcDpDcmVhdG9yVG9vbD5BZG9iZSBQaG90b3Nob3AgQ1MgTWFjaW50b3NoPC94YXA6Q3JlYXRvclRvb2w+CiA8L3JkZjpEZXNjcmlwdGlvbj4KCiA8cmRmOkRlc2NyaXB0aW9uIHJkZjphYm91dD0ndXVpZDo1MDFhZWI1OS1iZWRlLTExZGItOTg2MS05OWI4MDIyNDRkMjInCiAgeG1sbnM6eGFwTU09J2h0dHA6Ly9ucy5hZG9iZS5jb20veGFwLzEuMC9tbS8nPgogIDx4YXBNTTpEb2N1bWVudElEPmFkb2JlOmRvY2lkOnBob3Rvc2hvcDo1MDFhZWI1OC1iZWRlLTExZGItOTg2MS05OWI4MDIyNDRkMjI8L3hhcE1NOkRvY3VtZW50SUQ+CiA8L3JkZjpEZXNjcmlwdGlvbj4KCiA8cmRmOkRlc2NyaXB0aW9uIHJkZjphYm91dD0ndXVpZDo1MDFhZWI1OS1iZWRlLTExZGItOTg2MS05OWI4MDIyNDRkMjInCiAgeG1sbnM6ZGM9J2h0dHA6Ly9wdXJsLm9yZy9kYy9lbGVtZW50cy8xLjEvJz4KICA8ZGM6Zm9ybWF0PmltYWdlL2pwZWc8L2RjOmZvcm1hdD4KIDwvcmRmOkRlc2NyaXB0aW9uPgoKPC9yZGY6UkRGPgo8L3g6eG1wbWV0YT4KICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgIAogICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgCiAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAKICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgIAogICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgCiAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAKICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgIAogICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgCiAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAKICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgIAogICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgCiAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAKICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgIAogICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgCiAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAKICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgIAogICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgCiAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAKICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgIAogICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgCiAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAKICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgIAogICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgCiAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAKICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgIAogICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgCiAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAKICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgIAogICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgCiAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAKICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgIAogICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgCiAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAKICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgIAogICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgCiAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAKICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgIAogICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgCiAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAKICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgIAogICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgICAgCjw/eHBhY2tldCBlbmQ9J3cnPz7/4gxYSUNDX1BST0ZJTEUAAQEAAAxITGlubwIQAABtbnRyUkdCIFhZWiAHzgACAAkABgAxAABhY3NwTVNGVAAAAABJRUMgc1JHQgAAAAAAAAAAAAAAAAAA9tYAAQAAAADTLUhQICAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAABFjcHJ0AAABUAAAADNkZXNjAAABhAAAAGx3dHB0AAAB8AAAABRia3B0AAACBAAAABRyWFlaAAACGAAAABRnWFlaAAACLAAAABRiWFlaAAACQAAAABRkbW5kAAACVAAAAHBkbWRkAAACxAAAAIh2dWVkAAADTAAAAIZ2aWV3AAAD1AAAACRsdW1pAAAD+AAAABRtZWFzAAAEDAAAACR0ZWNoAAAEMAAAAAxyVFJDAAAEPAAACAxnVFJDAAAEPAAACAxiVFJDAAAEPAAACAx0ZXh0AAAAAENvcHlyaWdodCAoYykgMTk5OCBIZXdsZXR0LVBhY2thcmQgQ29tcGFueQAAZGVzYwAAAAAAAAASc1JHQiBJRUM2MTk2Ni0yLjEAAAAAAAAAAAAAABJzUkdCIElFQzYxOTY2LTIuMQAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAWFlaIAAAAAAAAPNRAAEAAAABFsxYWVogAAAAAAAAAAAAAAAAAAAAAFhZWiAAAAAAAABvogAAOPUAAAOQWFlaIAAAAAAAAGKZAAC3hQAAGNpYWVogAAAAAAAAJKAAAA+EAAC2z2Rlc2MAAAAAAAAAFklFQyBodHRwOi8vd3d3LmllYy5jaAAAAAAAAAAAAAAAFklFQyBodHRwOi8vd3d3LmllYy5jaAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAABkZXNjAAAAAAAAAC5JRUMgNjE5NjYtMi4xIERlZmF1bHQgUkdCIGNvbG91ciBzcGFjZSAtIHNSR0IAAAAAAAAAAAAAAC5JRUMgNjE5NjYtMi4xIERlZmF1bHQgUkdCIGNvbG91ciBzcGFjZSAtIHNSR0IAAAAAAAAAAAAAAAAAAAAAAAAAAAAAZGVzYwAAAAAAAAAsUmVmZXJlbmNlIFZpZXdpbmcgQ29uZGl0aW9uIGluIElFQzYxOTY2LTIuMQAAAAAAAAAAAAAALFJlZmVyZW5jZSBWaWV3aW5nIENvbmRpdGlvbiBpbiBJRUM2MTk2Ni0yLjEAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAHZpZXcAAAAAABOk/gAUXy4AEM8UAAPtzAAEEwsAA1yeAAAAAVhZWiAAAAAAAEwJVgBQAAAAVx/nbWVhcwAAAAAAAAABAAAAAAAAAAAAAAAAAAAAAAAAAo8AAAACc2lnIAAAAABDUlQgY3VydgAAAAAAAAQAAAAABQAKAA8AFAAZAB4AIwAoAC0AMgA3ADsAQABFAEoATwBUAFkAXgBjAGgAbQByAHcAfACBAIYAiwCQAJUAmgCfAKQAqQCuALIAtwC8AMEAxgDLANAA1QDbAOAA5QDrAPAA9gD7AQEBBwENARMBGQEfASUBKwEyATgBPgFFAUwBUgFZAWABZwFuAXUBfAGDAYsBkgGaAaEBqQGxAbkBwQHJAdEB2QHhAekB8gH6AgMCDAIUAh0CJgIvAjgCQQJLAlQCXQJnAnECegKEAo4CmAKiAqwCtgLBAssC1QLgAusC9QMAAwsDFgMhAy0DOANDA08DWgNmA3IDfgOKA5YDogOuA7oDxwPTA+AD7AP5BAYEEwQgBC0EOwRIBFUEYwRxBH4EjASaBKgEtgTEBNME4QTwBP4FDQUcBSsFOgVJBVgFZwV3BYYFlgWmBbUFxQXVBeUF9gYGBhYGJwY3BkgGWQZqBnsGjAadBq8GwAbRBuMG9QcHBxkHKwc9B08HYQd0B4YHmQesB78H0gflB/gICwgfCDIIRghaCG4IggiWCKoIvgjSCOcI+wkQCSUJOglPCWQJeQmPCaQJugnPCeUJ+woRCicKPQpUCmoKgQqYCq4KxQrcCvMLCwsiCzkLUQtpC4ALmAuwC8gL4Qv5DBIMKgxDDFwMdQyODKcMwAzZDPMNDQ0mDUANWg10DY4NqQ3DDd4N+A4TDi4OSQ5kDn8Omw62DtIO7g8JDyUPQQ9eD3oPlg+zD88P7BAJECYQQxBhEH4QmxC5ENcQ9RETETERTxFtEYwRqhHJEegSBxImEkUSZBKEEqMSwxLjEwMTIxNDE2MTgxOkE8UT5RQGFCcUSRRqFIsUrRTOFPAVEhU0FVYVeBWbFb0V4BYDFiYWSRZsFo8WshbWFvoXHRdBF2UXiReuF9IX9xgbGEAYZRiKGK8Y1Rj6GSAZRRlrGZEZtxndGgQaKhpRGncanhrFGuwbFBs7G2MbihuyG9ocAhwqHFIcexyjHMwc9R0eHUcdcB2ZHcMd7B4WHkAeah6UHr4e6R8THz4faR+UH78f6iAVIEEgbCCYIMQg8CEcIUghdSGhIc4h+yInIlUigiKvIt0jCiM4I2YjlCPCI/AkHyRNJHwkqyTaJQklOCVoJZclxyX3JicmVyaHJrcm6CcYJ0kneierJ9woDSg/KHEooijUKQYpOClrKZ0p0CoCKjUqaCqbKs8rAis2K2krnSvRLAUsOSxuLKIs1y0MLUEtdi2rLeEuFi5MLoIuty7uLyQvWi+RL8cv/jA1MGwwpDDbMRIxSjGCMbox8jIqMmMymzLUMw0zRjN/M7gz8TQrNGU0njTYNRM1TTWHNcI1/TY3NnI2rjbpNyQ3YDecN9c4FDhQOIw4yDkFOUI5fzm8Ofk6Njp0OrI67zstO2s7qjvoPCc8ZTykPOM9Ij1hPaE94D4gPmA+oD7gPyE/YT+iP+JAI0BkQKZA50EpQWpBrEHuQjBCckK1QvdDOkN9Q8BEA0RHRIpEzkUSRVVFmkXeRiJGZ0arRvBHNUd7R8BIBUhLSJFI10kdSWNJqUnwSjdKfUrESwxLU0uaS+JMKkxyTLpNAk1KTZNN3E4lTm5Ot08AT0lPk0/dUCdQcVC7UQZRUFGbUeZSMVJ8UsdTE1NfU6pT9lRCVI9U21UoVXVVwlYPVlxWqVb3V0RXklfgWC9YfVjLWRpZaVm4WgdaVlqmWvVbRVuVW+VcNVyGXNZdJ114XcleGl5sXr1fD19hX7NgBWBXYKpg/GFPYaJh9WJJYpxi8GNDY5dj62RAZJRk6WU9ZZJl52Y9ZpJm6Gc9Z5Nn6Wg/aJZo7GlDaZpp8WpIap9q92tPa6dr/2xXbK9tCG1gbbluEm5rbsRvHm94b9FwK3CGcOBxOnGVcfByS3KmcwFzXXO4dBR0cHTMdSh1hXXhdj52m3b4d1Z3s3gReG54zHkqeYl553pGeqV7BHtje8J8IXyBfOF9QX2hfgF+Yn7CfyN/hH/lgEeAqIEKgWuBzYIwgpKC9INXg7qEHYSAhOOFR4Wrhg6GcobXhzuHn4gEiGmIzokziZmJ/opkisqLMIuWi/yMY4zKjTGNmI3/jmaOzo82j56QBpBukNaRP5GokhGSepLjk02TtpQglIqU9JVflcmWNJaflwqXdZfgmEyYuJkkmZCZ/JpomtWbQpuvnByciZz3nWSd0p5Anq6fHZ+Ln/qgaaDYoUehtqImopajBqN2o+akVqTHpTilqaYapoum/adup+CoUqjEqTepqaocqo+rAqt1q+msXKzQrUStuK4trqGvFq+LsACwdbDqsWCx1rJLssKzOLOutCW0nLUTtYq2AbZ5tvC3aLfguFm40blKucK6O7q1uy67p7whvJu9Fb2Pvgq+hL7/v3q/9cBwwOzBZ8Hjwl/C28NYw9TEUcTOxUvFyMZGxsPHQce/yD3IvMk6ybnKOMq3yzbLtsw1zLXNNc21zjbOts83z7jQOdC60TzRvtI/0sHTRNPG1EnUy9VO1dHWVdbY11zX4Nhk2OjZbNnx2nba+9uA3AXcit0Q3ZbeHN6i3ynfr+A24L3hROHM4lPi2+Nj4+vkc+T85YTmDeaW5x/nqegy6LzpRunQ6lvq5etw6/vshu0R7ZzuKO6070DvzPBY8OXxcvH/8ozzGfOn9DT0wvVQ9d72bfb794r4Gfio+Tj5x/pX+uf7d/wH/Jj9Kf26/kv+3P9t////7gAOQWRvYmUAZEAAAAAB/9sAhAABAQEBAQEBAQEBAQEBAQEBAQEBAQEBAQEBAQEBAQEBAQEBAQEBAQEBAQEBAgICAgICAgICAgIDAwMDAwMDAwMDAQEBAQEBAQEBAQECAgECAgMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwP/wAARCAEEAMgDAREAAhEBAxEB/90ABAAZ/8QBogAAAAYCAwEAAAAAAAAAAAAABwgGBQQJAwoCAQALAQAABgMBAQEAAAAAAAAAAAAGBQQDBwIIAQkACgsQAAIBAwQBAwMCAwMDAgYJdQECAwQRBRIGIQcTIgAIMRRBMiMVCVFCFmEkMxdScYEYYpElQ6Gx8CY0cgoZwdE1J+FTNoLxkqJEVHNFRjdHYyhVVlcassLS4vJkg3SThGWjs8PT4yk4ZvN1Kjk6SElKWFlaZ2hpanZ3eHl6hYaHiImKlJWWl5iZmqSlpqeoqaq0tba3uLm6xMXGx8jJytTV1tfY2drk5ebn6Onq9PX29/j5+hEAAgEDAgQEAwUEBAQGBgVtAQIDEQQhEgUxBgAiE0FRBzJhFHEIQoEjkRVSoWIWMwmxJMHRQ3LwF+GCNCWSUxhjRPGisiY1GVQ2RWQnCnODk0Z0wtLi8lVldVY3hIWjs8PT4/MpGpSktMTU5PSVpbXF1eX1KEdXZjh2hpamtsbW5vZnd4eXp7fH1+f3SFhoeIiYqLjI2Oj4OUlZaXmJmam5ydnp+So6SlpqeoqaqrrK2ur6/9oADAMBAAIRAxEAPwDbGoqEekBbA/S31a9v+TffuvdKyjoDx6VNrf6y/Tnm3Pv3XulLS0Q4/wBh6rHjgcKOPqffuvdP0NL9LD/Gw5P/ABoe/de6dYaS/wCB/iPoBe35+v8AtvfuvdOkVIOLj/b8AW/wtzf/AGPv3XunKKn/AKC5/qRzf/D+nv3Xup0dK31II/2PP15/wHv3XupaUn50k/0IBI/3nj6+/de6kLScC4/4kk/7D37r3WcUv+0g/wC3Hv3Xusgprfgf7x9f9iffuvdc/tx/h/vv9h7917rv7a/9m3+sP+Ne/de64mlA/Fv9if8AiffuvdYjS/jT/sfz/vH5Pv3XusD0gP4/24B/4offuvdQ5KO34/2/H+2/1/fuvdQZaQj6L/X/AG3HFx6T7917qBLRg/ixte4H+v8A049+690zVFABc6QCSf8AgrH6/X8e/de6aJ6O1wV454PP+xU+/de6aJ6If0/w1Dix/wCNe/de6YamhIuQt73v/j+OR/X37r3SYrMepDELdT/S9x/r8X+o/wBf37r3X//Q2+aKh+lh/wAGe3NjwQv+29+690p6alAsAth+FAvf/E/19+690/U9Na3+2vbgf8FH9ffuvdPUFL9OP6cDkn6f7H/Ye/de6d4KS9vT/wARb/YWsOffuvdOsNF9Lj/euL2Frm4sL+/de6dIqRQB/rD6WH0/x5v/ALD37r3UxKcf2R/Xn8/7zc+/de6krT3/AB/vv9jc/wC8e/de6zrTcfgf6/8AxHv3Xusn26j8/wC8D37r3XMU4P8AU/42P/Fbe/de67+2/wAP99/yV7917rowEf2Qf9b/AIr+ffuvdcTEfytv9sPfuvdcDCD9U/2x5/3g+/de6xNTj/H8fqH+39+691gaD/D/AH3+8g+/de6iSUotwP8AYW4/2/0uPfuvdN81Hfi39OD/AK1/ryCffuvdNk1LYNxf+oPI+nNxY3v/ALb37r3TNUUd72Uj6nT/AE4vdT/xHv3XumWWl+vA/wBtY8/6of4e/de6Z6ij+tlNv6fQgfS4v9ffuvdJ2qoQbkCzG/P4b/Br/T37r3X/0dySmpf02H+t/vV+fwPfuvdPtPTfTj/C/wDU/wBF/wAPfuvdP9NSXtxbkcfX68D6Wv8AT37r3T9T0f04+n1/r9fyfwP9b6+/de6d4qUL+ORb8cfj8f8AEn37r3TjHB/h/T/ff4/717917qbHD/Xk/wDFffuvdSliH9Ln82H/ABTn37r3WdYWP0AX/X/P+2uffuvdZFpyf7X+2H/Ekj37r3WYU4H/ABUkf8b9+691y8H+P+8/8a9+6910YT+D/vv949+69114SPrf/YD/AJH7917rrQB9b+/de66MYP8AxsX9+691jMX+H+29+691gaEc2/23++49+691FaEj/Dj/ABP+3/I9+691FeIH6ix9+691Bmpxc8f8U549+6900z0tr8X/AMf6/wCsfwffuvdMtRS35Auf9a17/gj+tvfuvdMs9NYfQ2/3kH/X9+690w1VJa5tx+ePp/iP8Pfuvdf/0t0GmpuRx/r/AFtb8AfTj37r3SjpKMWBI/2FiLD6fj6e/de6UdNSHhitl+ot+b/0/oPfuvdPUUFha1rf4f77/b/X37r3U5Ih+AOD9f8AfE8+/de6mRxX/H++/qf6D37r3UtY/pxf/eh7917qSkf9Bf8Ax+gHv3XuswQfnn/evfuvdc/fuvddhSfx7917rnoP5P8AxP8AxT37r3Xegf1Pv3Xuu9C/77/e/fuvdeKKfx7917rG0Kn6Eg/7x7917rGYSPz/ALED/jfv3XusRQ/kX/334/Pv3XusTRA/4f4H/fX9+691EeK31H+x+n1/1vfuvdRHht9OR/vv6/T37r3UGWAEGwHH1U2/5Fx7917pnqKYHkD/AB5H+8H/AB/x9+690yT0/wBeP8PpwRzxa319+690x1FP9eP9Y/0+nB/r/h7917r/092OkpPobX+lh+COPUf6Dn37r3SppaQAAsL/AJsfx9OW/qT/AE9+6909RRAfj/XP+++p9+691MRL8AG3v3XupscX0/w/23+w/qffuvdTI4ibcWX+v+++vv3XupSoqj6X/wBf37r3XP37r3XMIfzx7917rIFA/wCK/n37r3Xfv3Xuve/de697917r3v3Xuve/de697917r3v3XuuJUH/X/r7917rEyW+oB/x9+691haO/0/23v3XuokkVvxb/AGHB9+691Cli/Nrf63+9/wCI9+691Amhv+Bf+v8AX37r3TLU0978f8VN/wDeOP8AX9+690yT0/8AgTb/AA+vHII/Pv3Xuv/U3iKOmsA5Bv8AW1vr/T/be/de6fIY/wDit/8Aif8AYe/de6non0/3gf1/1/fuvdTY0J4H+F/fuvdOEcIABYf8g/8AFffuvdSPfuvddhSfp/t/x7917rKFA/1/6+/de65Wv9Pfuvdcwh/1vfuvdd+P/H/ePfuvdd6B/j7917rlpH9B7917r2kf0H+29+691xKD/Ee/de64lCPpz7917rh7917r3v3Xuve/de6xsn5H+29+691iZQeGH++/2Pv3XuockVv8R+LfX/Ye/de6gSR2J4uD7917pvnhvyLf8VHH1/xHv3XumOog+p/P1/25+tj9Pr7917r/1d56GL6H/ff7b+p9+6905Rp+P9vb/D37r3U+KMn8f8a/wH+w/wBt7917p0jjCD/H/ev9b37r3WX37r3XJVvz+P8AffT37r3WcL+APp7917rIEA+vP+9e/de65+/de697917r3v3Xuve/de697917r3v3Xuve/de697917roqD9f9v7917rEVI/1v6+/de64+/de697917rogEe/de6wMv4Pv3Xuocsdv9b/fcj+nv3Xum6WK3+t/vgD7917pqnivf/Y/7Dg/09+691//1t6yNLAf1P0/2P59+6904Qx/T/ff77/D37r3TrDHpF/9sPfuvdZ/fuvdclW5/wAPz/xT37r3UhVv/gP999PfuvdZQABYe/de679+691737r3Xvfuvde9+691737r3Xvfuvde9+691737r3Xvfuvde9+691737r3XAoD/AIf71/tvfuvdYyCPr/t/x7917rr37r3XFlv/AK/v3Xuo7LcW/P8AvuPfuvdQpE/H+2/437917ptmj4vbi1iD7917r//X3tYl1G/4H++/3n37r3TtBH9CR/xr/ilvfuvdTffuvddgEn37r3UlE/2w/wB59+691m9+691737r3Xvfuvde9+691737r3Xvfuvde9+691737r3Xvfuvde9+691737r3Xvfuvde9+691737r3XvfuvdYmW3I+n+9e/de64e/de64Ov5H+x/33+Hv3XuokiX/2P+8H/jfv3XuoEi/X+h4P+B9+691//9De/hS+kf1+v+sPfuvdOyCw/wBf/eh9Pfuvdc/fuvdSET8f7c+/de6ke/de697917r3v3Xuve/de697917r3v3Xuve/de697917rsC/v3Xuu9J/335/3n6e/de69pPP++/41b37r3XRBH19+69117917r3v3Xuve/de697917r3v3XusTLbkfT/AHr37r3XD37r3WB1tx+D9P8Aff4e/de6gyL/ALzwf9f37r3X/9HfKgT8/wCq4H+sDz7917qd7917rIg/P+wHv3XupSiw/wAfz/vv8Pfuvdcvfuvde9+691737r3Xvfuvde9+691737r3XYBP9ffuvdZNPP8ATnj/AGB/4n37r3XIC1/8ffuvdd+/de697917rogHi3v3XuuDL/T/AA/2P0HPP/Ee/de6x+/de697917r3v3Xuve/de697917rEy25H0/3r37r3WJhcH+v4/33+Pv3Xuoki3H+v8A7x/T8e/de6//0t9OJbD/AFgB7917rL7917qVGv0/oP8Ae/fuvdZvfuvde9+691737r3Xvfuvde9+691737r3WRV/r/X/AA/3ke/de6ye/de697917r3v3Xuve/de697917r3v3Xuve/de6xsAAP9744/4rf37r3WP37r3Xvfuvde9+691737r3XvfuvdYmW3I+n+9e/de6hyr9f9uPfuvdf/099ZP0/6/wDyL37r3WRRdh/hz/vv9j7917qWg4/1/fuvdc/fuvde9+691737r3Xvfuvde9+691zQXP8ArW/3n/io9+691l+nv3Xuve/de697917r3v3Xuve/de697917r3v3Xuve/de68eR/vv8AjXv3XusB+v8AvPv3Xuuvfuvde9+691737r3XvfuvdeIuLe/de6hyD/iQf99+ffuvdf/U31wLAD37r3WRPr/sP+JHv3Xupg4AH+Hv3Xuu/fuvde9+691737r3Xvfuvde9+691mX6f7Af8a59+691y9+691737r3Xvfuvde9+691737r3Xvfuvde9+691737r3XvfuvdYW+v5/2Pv3XuuPv3Xuve/de697917r3v3Xuve/de6jSj9X+vf/AG//ACP37r3X/9XfY9+691zj/V/sP+JHv3Xupvv3Xuve/de697917r3v3Xuve/de67X6j/XH+9+/de6zDgD/AH3/ABX37r3Xfv3Xuve/de697917r3v3Xuve/de697917r3v3Xuve/de697917rE/wDvv9jf/H8Ae/de64e/de697917r3v3Xuve/de697917qPL/a/2H/Ee/de6/9bfY9+691zj+v8AsP8AiR7917qb7917r3v3Xuve/de697917r3v3Xuux/j/AEP+9cf7z7917rMPp/T/AA9+69137917r3v3Xuve/de697917r3v3Xuve/de697917r3v3Xuve/de6xPa/8Atv8AbH/Yn37r3XD37r3Xvfuvde9+691737r3XvfuvdYJP7X+t/xHv3Xuv//X32Pfuvdc0+v+w/4ke/de6m+/de697917r3v3Xuve/de697917r3v3Xusqnj/AH1ubcc8ce/de65+/de697917r3v3Xuve/de697917r3v3Xuve/de697917r3v3XusLfX+nv3XuuPv3Xuve/de697917r3v3Xuve/de6jy/2v9h/xA9+691//9DfY9+691yT9Q9+691MU3Uf7b/be/de65e/de697917r3v3Xuve/de697917rsG3+3/ANh/T/b+/de6yhr/APEfXn/eB7917rl7917r3v3Xuve/de697917r3v3Xuve/de697917rixtf8A1v8AbfXn37r3WH37r3Xvfuvde9+691737r3Xvfuvde9+691EkPpJ/qf+N+/de6//0d9WNtSg/wCw/wBt7917rIDYg/4+/de6lofqP9j7917rJ7917r3v3Xuve/de697917r3v3Xuve/de697917rJr55tx/r8n/jfv3XuuYP+I/3j/Y/k+/de67Bv7917r3v3Xuve/de697917rxNuT7917rCzXP+H49+691x9+691737r3Xvfuvde9+691737r3XTGwPv3XuoUx4t/vuf8AjXv3Xuv/0t9CBvqv+xHv3XupHv3Xus6E2B/P++H+8+/de6k/Xke/de697917r3v3Xuve/de697917r3v3Xuve/de697917r3v3XuuwSP6/7c8f63I9+691y1/wCw/qfr9fzbj8+/de671W/N/wDXvcn/AHng+/de64lj/wAbtb/be/de64+/de697917r3v3Xuve/de697917r3v3Xuve/de6xufoP8AY+/de6gStc/X+p/4p/vHv3Xuv//T3yUblWH9Rf8A437917qeDcX9+691kQ/Uf7H37r3UhD+D/sP99/j7917rJ7917r3v3Xuve/de697917r3v3Xuve/de697917r3v3Xuve/de697917r3v3Xuve/de697917r3v3Xuve/de697917r3v3Xuve/de697917qNI31P9eP9h/yL37r3TdI31/x/wB6/wCN+/de6//U3v6eS62J5H++H+8e/de6dI2+gvwfp/r+/de6zA2IPv3Xus4Nxf37r3WZWv8AX6/737917rl7917r3v3Xuve/de697917r3v3Xuve/de697917r3v3Xuve/de697917r3v3Xuve/de697917r3v3Xuve/de697917r3v3XuuDmwt/t/8AW9+691Alf6/X/jQ/4k+/de6gsfqT/r/8a9+691//1d7CCTSQR/T/AG6n8i39PfuvdPML6h9f99/xv37r3U1WuLfkf76/v3XusyN+D/sP+Ke/de6ye/de6yq9+D9f97/437917rn7917r3v3Xuve/de697917r3v3Xuve/de697917r3v3Xuve/de697917r3v3Xuve/de697917r3v3Xuve/de66JAFz7917qHI97/0/P+P+A9+691Cka5PH/FPp/vdvfuvdRJGsLf76/wDT37r3X//W3n6eb6An/W/3nj/WP49+6908wSkEC/H4Pv3XunWNwwBB5/33P+x9+691KBuAffuvdZ1a/wDrj37r3XL37r3XMOR9eR/vPv3XusgYH6f7b8+/de679+691737r3Xvfuvde9+691737r3Xvfuvde9+691737r3Xvfuvde9+691737r3Xvfuvde9+691FllB4B4/wB9/T6+/de6gyP/ALb8D/ib+/de6jM2kX/23v3XuoMj/Uk8D/ef8ffuvdf/194yCUG3P+H/ABX/AG3v3XunmCb6XPIt/sf6Efj37r3TtBPbkH/in+2/p7917p2hkVh9RzY2/of+Ke/de6z/AE5Hv3Xus4YN7917rv37r3Xvfuvdcg7D/H/X/wB9f37r3XNXvweD/vfv3Xuufv3Xuve/de697917r3v3Xuve/de697917r3v3Xuve/de697917riWC/8U/3309+691Dkmvxc/wCt/vvp7917qKz/AJP+wHv3Xuo7N+T/ALAe/de6hyPe/P8Ar/8AFPfuvdNVVUWFlP0PA/1X9f8AYD37r3X/0N3amn+hHFrA3/r9Pp/Q+/de6e4Jgbf1/p9bH6f7Y+/de6doZfp/vufz/sf979+6905RTlbc2+nH+8/7H37r3TrHPcf8V/3j/Wv7917qUG/IP/GvfuvdZla4/wAfz7917rl7917r3v3Xuve/de6yK/4P+3/4r7917rJ7917r3v3Xuve/de697917r3v3Xuve/de697917rG8gH0P++/w/r7917qFJKT9P+JP/I/fuvdRWe305P8Avv8Ab+/de6wM/wDsT7917qO72/PP+P4/Pv3Xum2onVQRe9vqB9b/AEt/sffuvdME89rsTdvoB+f8AP62v7917r//0d1ajrNQHqAJBtf/AGPI54AA9+690o6eovax/H5t/tj9B7917p6gnBt/U2/2Nv6c/X37r3TnFL/j/vX++v7917qfFMV+nI/3r/ePfuvdOEc/0sR/X/kf5+vv3XupyTqfzY/1/H/FPfuvdSRJ/X/bj37r3WT68j37r3Xvfuvde9+6912CR9PfuvdZA4/PHv3Xuudwfob+/de697917r3v3XuuDSKPz7917qNJP/vv+Nf8V9+691Dea9/z/vv6+/de6jtJ/U/7Ae/de6wlyf8AD/W9+691hdwoP+H1P4Hv3Xum2oqQoNj/ALz+f6f4n37r3TFUVPJub2+g/oP6f0v7917pmmnvcsbKP9t/rD/fc+/de6//0ty2kqvpzyPr/jf6MPwffuvdKmkrCbBjzwLjgH6Acf6on37r3SggqePr/sD9b/0/wt7917p4hqfoNXBt/S4/1/8AW9+6905Rz/Tm/wDvv68/7z7917qdHN9P98L/APEe/de6lpN/Q3/2Nj/vHv3XupSVDDgE/wCt9P8AeOR7917qUtUfz/xTj/evfuvdZ1qQfr/sPp/xFvfuvdZBOptyvP4N/wDivv3XuuXl/wCC/wC+/wAffuvdcvKv5/3sf8a9+6910ZlH0/33+2v7917riZz/AK/+3/4k+/de64Gdv8B7917qO8/+P+x+n/Gz7917qO0pP++49+691iLE/n6/j37r3XAuq/U/T37r3UZ57C/0H9f8fpx7917psnqgL82+o/Nzf/Ae/de6ZKirv+f8OPxf8ce/de6aZZ/yTb/jf9ffuvdMdXVg/Rvof99b8E/7f37r3X//09wairQdPqBJtb/UsPpwb/U/09+690qKWrvb1fT+v+vbkf0P9ffuvdKOlrT9Cbj/ABPqH0B0m/P19+690+09WDY3vxyfoeTyCPqD7917p2hqPob82ufyLfi9uD7917pyiqRwCSP96/r9f6H37r3U+OYHkG3+t9P95+nv3XupSzH88/T/AH3++v7917qQs4/xI/2H/I/fuvdZVmX/AFRX825t/wAU9+691mEoP5U/6x9+691yEqj82/1iP+K+/de65ef+hH/Jo/4r7917riakj+n+9/71b37r3XA1Lfj+v+t/xu3Pv3XuuBmck/T/AH3+N/fuvdcPIf8AD/ef+K+/de668h/w/wB54/1+ePfuvdY2l/x/3mw/4pz7917qJLUKo5YE/wBL/T/iPz7917prnq/xf/X54/43z+PfuvdM01SSTY2/xHBI+v8AsF9+6901TVAW/P0/P4H9QB7917pjqq08gE/8T/hc/i/v3Xuk9U1lwfVYcC/4vf6Ac39+691//9TbPocgOB/Xkpe3q/qp/BuPp/T/AG/v3XulZRZEceskACx/tp/gw/IH+x9+690pqavBAOoWI4N/SfyP8UPPv3Xun+nruRcm4/x9QtybngMPfuvdP1PkBxc8A8n/AG3Okc8+/de6eoKoMOGBH+uLW/3n37r3TjHU3tZr/wCsfpz/AL37917qbHUt/X/Yf8VH19+691KWo+n9f8Dbj/W5Pv3Xusyzg/1/2Njb/bH37r3WQTD/AJFcf8R7917rmJT/AFNv8Df37r3XISX/ALR/2JPv3XuuWsj+1/t/+N+/de695R/h/vP++Pv3XuveYf4f7Y+/de64NPb+gH+NgP8Aeb+/de6wPVIv1YH/AA5b+v8ATj37r3UCWuvwpsP8bXP+Nhx/vfv3XumyWrve5uP8TxwOOf8AW9+6901S1Yvyef6n/iALk+/de6a5621xf/Yfk/T6gXtb37r3THUVl7ksP9v9OP68/Ue/de6YaqsHNzYf7Yn+oFjzce/de6TtVW/W5B+tl4OmxNr3uOf8Pr7917r/1dpejr+RpJDD1FL88f6lv6X9+690qKPJXANyCBY2PrH/AAYA+of77n37r3SmpMlax1Wv+RfQ3FvUoP6v969+690pKXJXsNQX6W5vGbH6Kf7PPv3Xun+nyH05IJ5F/p+DwRa44/w9+6909wVxBHqKn8MDwTc25H+Hv3XunmHJHgk3/wAf6/69ufp/U+/de6dYsiLDk2/r+oDj+os1vfuvdOCVy8cgfi5YfW304P1/3v37r3UtatSL/Xn8W/3u4A9+691IWqTjn6/1BP8AxT37r3WUVK/UMPr9QeP+J9+691yFUL/qBt/vv6c29+6912apAOWtfj6D/eD7917rEa5R9D9b/wCNrf6w+h9+691ieu/oT/0L/trc/wC8+/de6hPXD8Ek3/sj/D63P19+691EerP1+n49Rt/vJJ59+691DkrODY3/AMT+n/iPp7917pumq/qWf/iB/rAfn37r3TVPXWvY2vcX+rf7D8D/AGPv3XumWeu+ov8AX/EkkH63Jt+ffuvdMtRWgXAOpufSD/Qf2j/sPfuvdJ6qriC3quRfUbgIpA5sPzp/x9+690l62vuDZ7L9SxuL3/x+tr/7E+/de6//1tl+kyPpUgh0BGl1OlkAsLW+tx/vjyPfuvdKilyAOn1X54kU8i9iAyjm/wDX/evfuvdKKlyBFizfW1nH0It9HX/fW/w9+690pKXIDizAE2v+Q3+wN+PfuvdKGmyFuCbXP0J9P+vz9CffuvdP9PkeANV/8L2/3v6j37r3TxBkBxZh/rXt9D/Qmx59+6906xV309XPHI4IPP44v7917qdHXNcWYk/6/It/gR9R7917qWtefy31FvVe/wCR9eQPr/vPv3XupSVzcDUfofobjn/AH37r3Uha9voWJ4/JNvx+OeffuvdchXn/AFQ+t7m9z/vXv3XuuzkP9Y8c/wDFOW59+691wNcT+QP8b/43txz7917rG1WefV+Afr/QfXk/4+/de6wPW/T12/1uP96A9+691FkrR/qiT/sb+/de6gTV/wBfUB/iTc8cfQe/de6a5q/68/05b6/X8KLkj37r3TRPW/W7f8hfUn8cADj37r3TTUV1gTq0j/k9hY3t/qf99z7917pP1eQAFr6R+VuSWP5JIHN7+/de6TlXkABd30LzZb+puR9FvyAf9t7917pJV2S1XudK86UBuWIJFz+D9ffuvdf/19jKj8+seK2uw1Wtp/F/1cab/wCx9+690pqX7i40/wCc0nXa9v8AG3+Gr/effuvdKik+5sLWvzqt+n/efz/vHv3XulBS/cXH1t+fra1mtb8/7b37r3Sjpvurrq/Tx9fra39m3Nre/de6fab7j/Yfi/8AW3Fv8L+/de6eoPuPx9OP1fX/AGH+N/fuvdOsX3Pp+trfi9vx9P8AH6+/de6dI/vOL6bX512v/wAm8+/de6mp9xx9NP50a7X/AML8fX37r3UlfNcfq/2F7/n6W/Pv3XupCea39u3/ACFf6f4e/de6yjy/nVb/AGq//I/fuvddfvf7R/yf/T/inv3XuuH79ze31P8AT/eLc/8AE+/de64N5v8AD/HVr9+691HPntzf/Yab/wDJvPv3XuoUn3P4ta/+P+w+v59+6902y/cc6bfXm36r/n6/4+/de6bJfNc/XV+b/W3P9fz7917ptqPuLHRbVx9barW/Gr03/p7917piqfuedX6vTfX9PoPrp5+nv3XumCo+4sbW8luNf0/H0tx9P959+690la37rU3k/X/j/W3pvb8W/p7917pNT+a5v+u4vqv9L/2dPH/Ee/de6//Z"
+    open("Blank.jpg", "wb").write(base64.b64decode(img))
+    DeepFace.analyze("Blank.jpg", actions=["emotion"], enforce_detection=False)
+    remove("Blank.jpg")
+
+
+def analyze_frame(img, debug=False):
+    from deepface import DeepFace
+
+    try:
+        res = DeepFace.analyze(img, actions=["emotion"])
+        if debug:
+            frame = cv2.imread(img)
+            dominant_emotion = res[0]["dominant_emotion"]
+            debug_image = cv2.putText(
+                frame,
+                f"EMOTION: {dominant_emotion}",
+                (20, 20),
+                cv2.FONT_HERSHEY_SIMPLEX,
+                0.5,
+                (255, 0, 0),
+                1,
+                cv2.LINE_AA,
+            )
+            cv2.imwrite(f"image_{datetime.now()}.png", debug_image)
+        return res
+    except Exception as e:
+        if debug:
+            print(f"[PYTHON]: Error during analysis\nERROR: {e}")
+            cv2.imwrite(f"failed_{datetime.now()}.png", img)
+        raise Exception("Failed to detect face")
